requirements.txt
openai-whisper==20231117
torch
torchaudio
